+++ include/boost/math/ccmath/ccmath.hpp
@@ -1,0 +1,12 @@
+//  (C) Copyright Matt Borland 2021.
+//  Use, modification and distribution are subject to the
+//  Boost Software License, Version 1.0. (See accompanying file
+//  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+//
+
+#ifndef BOOST_MATH_CCMATH
+#define BOOST_MATH_CCMATH
+
+#include <boost/math/ccmath/sqrt.hpp>
+
+#endif // BOOST_MATH_CCMATH
