+++ test/sqrt_test.cpp
@@ -5,7 +5,7 @@
 
 #include <cmath>
 #include <limits>
-#include <boost/math/special_functions/sqrt.hpp>
+#include <boost/math/ccmath/sqrt.hpp>
 #include <boost/core/lightweight_test.hpp>
 
 template <typename Real>
@@ -15,7 +15,7 @@ void test_float_sqrt()
     
     constexpr Real tol = 2*std::numeric_limits<Real>::epsilon();
     
-    constexpr Real test_val = boost::math::sqrt(Real(2));
+    constexpr Real test_val = boost::math::ccmath::sqrt(Real(2));
     constexpr Real dummy = 1;
     static_assert(test_val > dummy, "Not constexpr");
 
@@ -23,17 +23,17 @@ void test_float_sqrt()
     BOOST_TEST(abs(test_val - known_val) < tol);
 
     // 1000 eps
-    constexpr Real test_1000 = boost::math::sqrt(1000*std::numeric_limits<Real>::epsilon());
+    constexpr Real test_1000 = boost::math::ccmath::sqrt(1000*std::numeric_limits<Real>::epsilon());
     Real known_1000 = std::sqrt(1000*std::numeric_limits<Real>::epsilon());
     BOOST_TEST(abs(test_1000 - known_1000) < tol);
 
     // inf
-    constexpr Real test_inf = boost::math::sqrt(std::numeric_limits<Real>::infinity());
+    constexpr Real test_inf = boost::math::ccmath::sqrt(std::numeric_limits<Real>::infinity());
     Real known_inf = std::sqrt(std::numeric_limits<Real>::infinity());
     BOOST_TEST_EQ(test_inf, known_inf);
 
     // NAN
-    constexpr Real test_nan = boost::math::sqrt(std::numeric_limits<Real>::quiet_NaN());
+    constexpr Real test_nan = boost::math::ccmath::sqrt(std::numeric_limits<Real>::quiet_NaN());
     Real known_nan = std::sqrt(std::numeric_limits<Real>::quiet_NaN());
     BOOST_TEST(std::isnan(test_nan) && std::isnan(known_nan));
 }
@@ -45,7 +45,7 @@ void test_int_sqrt()
 
     constexpr double tol = 2*std::numeric_limits<double>::epsilon();
 
-    constexpr double test_val = boost::math::sqrt(Z(2));
+    constexpr double test_val = boost::math::ccmath::sqrt(Z(2));
     constexpr double dummy = 1;
     static_assert(test_val > dummy, "Not constexpr");
